Fixed AStar2D test Fixed wrong coordinates used in texture rendering. Fixed input manager handling modifier keys incorrectly. Fixed more stuff I forgot.

+++ TLib/Files.hpp
@@ -37,7 +37,7 @@ static Path saveFileDialog(
     char* saveFilePath = tinyfd_saveFileDialog(
         title.c_str(),
         defaultPath.string().c_str(),
-        cfilters.size(),
+   (int)cfilters.size(),
         cfilters.data(),
         filterDescription.empty() ? NULL : filterDescription.c_str());
 
@@ -66,7 +66,7 @@ static Path openSingleFileDialog(
     char* openedFilePath = tinyfd_openFileDialog(
         title.c_str(),
         defaultPath.string().c_str(),
-        cfilters.size(),
+   (int)cfilters.size(),
         cfilters.data(),
         filterDescription.empty() ? NULL : filterDescription.c_str(), 0);
